Refactor to use the bloc pattern

+++ lib/weather.dart
@@ -151,13 +151,13 @@ class Main {
 
 class Wind {
   double speed;
-  int deg;
+  double deg;
 
   Wind({this.speed, this.deg});
 
   Wind.fromJson(Map<String, dynamic> json) {
-    speed = json['speed'];
-    deg = json['deg'];
+    speed = json['speed'].toDouble();
+    deg = json['deg'].toDouble();
   }
 
   Map<String, dynamic> toJson() {
